refactor: post/project List 컴포넌트 생성, types폴더/파일 생성
