+++ src/app/api.ts
@@ -1,27 +1,5 @@
 import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
-
-interface Post {
-  id: number;
-  title: string;
-  content: string;
-  author: string;
-  likes: number;
-}
-
-interface Project {
-  id: number;
-  name: string;
-  content: string;
-  field: string[];
-  createDate: string;
-  endDate: string;
-  likes: number;
-}
-
-interface User {
-  likedPosts: number[];
-  likedProjects: number[];
-}
+import { Post, Project, User } from "types";
 
 export const api = createApi({
   baseQuery: fetchBaseQuery({
