+++ src/pages/Main.tsx
@@ -1,20 +1,20 @@
 import { Button, Group, Stack } from "@mantine/core";
 import Layout from "components/Layout";
-import PostCard from "components/PostCard";
-import { Category } from "components/Category";
+import { MainCategory } from "components/MainCategory";
 import { api } from "app/api";
 import { Link } from "react-router-dom";
-import ProjectCard from "components/ProjectCard";
+import PostList from "components/post/PostList";
+import ProjectList from "components/project/ProjectList";
 
 export default function Main() {
   const { data: postsData, isFetching: isPostsFetching } =
     api.useGetAllPostsQuery(null);
-  const { data: ProjectsData, isFetching: isProjectsFetching } =
+  const { data: projectsData, isFetching: isProjectsFetching } =
     api.useGetAllProjectsQuery(null);
 
-  // if (isPostsFetching || isProjectsFetching) return <div>loading...</div>;
+  if (isPostsFetching || isProjectsFetching) return <div>loading...</div>;
   return (
-    <Layout navbarChildren={<Category />}>
+    <Layout navbarChildren={<MainCategory />}>
       <Stack
         mih={300}
         sx={(theme) => ({
@@ -32,12 +32,8 @@ export default function Main() {
           </Link>
         </Group>
 
-        {postsData?.data.map((post, i) => (
-          <PostCard key={i} {...post}></PostCard>
-        ))}
-        {ProjectsData?.data.map((project, i) => (
-          <ProjectCard key={i} {...project}></ProjectCard>
-        ))}
+        <PostList posts={postsData?.data} />
+        <ProjectList projects={projectsData?.data} />
       </Stack>
     </Layout>
   );
